Code cleanup

+++ src/main/java/network/crypta/node/simulator/LongTermMHKTest.java
@@ -65,8 +65,8 @@ public class LongTermMHKTest extends LongTermTest {
     System.out.println("DATE:" + dateFormat.format(today.getTime()));
     csvLine.add(dateFormat.format(today.getTime()));
 
-    System.out.println("Version:" + (int) Version.currentBuildNumber());
-    csvLine.add(String.valueOf((int) Version.currentBuildNumber()));
+    System.out.println("Version:" + Version.currentBuildNumber());
+    csvLine.add(String.valueOf(Version.currentBuildNumber()));
 
     int exitCode = 0;
     Node node = null;
